requirements.txt
pandas>=1.5.0
numpy>=1.21.0
rapidfuzz>=2.0.0
streamlit>=1.20.0
plotly>=5.0.0
PyYAML>=6.0
duckdb>=0.9.0
joblib>=1.3.0
